@@ -2,52 +2,35 @@ package chessengine.CentralControlComponents;
 
 import chessengine.ChessRepresentations.ClientsideChessGameWrapper;
 import chessserver.ChessRepresentations.ChessGame;
-import chessserver.Enums.CampaignTier;
+import chessserver.Enums.CampaignAttempt;
+import chessserver.Enums.CampaignLevel;
 import chessserver.Enums.Gametype;
 
-import java.util.Objects;
-
 public class ChessGameHandler {
-    private final ChessCentralControl control;
-    public ClientsideChessGameWrapper gameWrapper;
+    private boolean isCurrentGameFirstSetup;
+    // for campaign mode only
+    private CampaignAttempt campaignAttempt;
+    //
 
-    public boolean currentlyGameActive(){
+    public ClientsideChessGameWrapper gameWrapper;
+    public ChessGameHandler(ChessCentralControl control) {
+        gameWrapper = new ClientsideChessGameWrapper(control);
+    }
+
+    public boolean shouldSaveGame(){
+        return isActiveGame() && isCurrentGameFirstSetup() && !gameWrapper.getGame().isEmpty();
+    }
+
+    public boolean isActiveGame(){
         return gameWrapper.getGame() != null;
     }
-    // for campaign mode only
-    private int gameDifficulty;
-    private int levelOfCampaignTier;
-    private CampaignTier campaignTier;
-    private boolean isCurrentGameFirstSetup;
 
-    public ChessGameHandler(ChessCentralControl control) {
-        this.control = control;
-        gameWrapper = new ClientsideChessGameWrapper(control);
-
+    public CampaignAttempt getCampaignAttempt() {
+        return campaignAttempt;
     }
 
-    public int getLevelOfCampaignTier() {
-        return levelOfCampaignTier;
-    }
-
-    public void setLevelOfCampaignTier(int levelOfCampaignTier) {
-        this.levelOfCampaignTier = levelOfCampaignTier;
-    }
-
-    public CampaignTier getCampaignTier() {
-        return campaignTier;
-    }
-
-    public void setCampaignTier(CampaignTier campaignTier) {
-        this.campaignTier = campaignTier;
-    }
-
-    public int getGameDifficulty() {
-        return this.gameDifficulty;
-    }
-
-    public void setGameDifficulty(int gameDifficulty) {
-        this.gameDifficulty = gameDifficulty;
+    public void setCampaignAttempt(CampaignAttempt campaignAttempt) {
+        this.campaignAttempt = campaignAttempt;
     }
 
     public boolean isCurrentGameFirstSetup() {
@@ -55,31 +38,28 @@ public class ChessGameHandler {
     }
 
     public void clearGame() {
-        campaignTier = null;
-        levelOfCampaignTier = -1;
-        gameDifficulty = -1;
+        campaignAttempt = null;
         gameWrapper.clearGame();
     }
 
 
-    public void switchToNewGame(ChessGame newGame) {
-        setUpGameGui(newGame,false,null, true);
-
+    public void switchToNewGame(ChessGame newGame, boolean isVsComputer) {
+        setUpGameGui(newGame, isVsComputer, false,null, true);
     }
 
-    public void switchToGame(ChessGame newGame, boolean isFirstLoad) {
-        setUpGameGui(newGame,false,null, isFirstLoad);
+    public void switchToGame(ChessGame newGame, boolean isVsComputer,  boolean isFirstLoad) {
+        setUpGameGui(newGame, isVsComputer,false,null, isFirstLoad);
 
     }
     public void switchToOnlineGame(ChessGame newGame, Gametype gametype, boolean isFirstLoad) {
-        setUpGameGui(newGame,true,gametype, isFirstLoad);
+        setUpGameGui(newGame, false, true, gametype, isFirstLoad);
 
     }
 
 
-    private void setUpGameGui(ChessGame newSetup, boolean isWebGame, Gametype gametype, boolean isFirstSave) {
+    private void setUpGameGui(ChessGame newSetup, boolean isVsComputer, boolean isWebGame, Gametype gametype, boolean isFirstSave) {
         isCurrentGameFirstSetup = isFirstSave;
-        gameWrapper.loadInNewGame(newSetup,isWebGame,gametype);
+        gameWrapper.loadInNewGame(newSetup, isVsComputer, isWebGame, gametype);
 
 
 
