@@ -1,242 +1,185 @@
 package chessengine.Crypto;
 
-import chessserver.ChessRepresentations.ChessGame;
-import chessserver.User.FrontendClient;
+import chessengine.App;
+import chessserver.ChessRepresentations.GameInfo;
+import chessserver.Misc.ChessConstants;
+import chessserver.Misc.SavePath;
 import chessserver.User.UserInfo;
 import chessserver.User.UserPreferences;
+import chessserver.User.UserWGames;
+import com.fasterxml.jackson.core.type.TypeReference;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import org.nd4j.shade.jackson.databind.ObjectMapper;
+import com.fasterxml.jackson.databind.ObjectMapper;
 
-import javax.crypto.*;
+import javax.annotation.Nullable;
 import java.io.*;
+import java.nio.file.Files;
+import java.nio.file.Path;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.UUID;
+import java.util.concurrent.TimeUnit;
+import java.util.function.Consumer;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
+import java.util.concurrent.locks.ReentrantReadWriteLock;
+
 
 
 public class PersistentSaveManager {
     private final static Logger logger = LogManager.getLogger("Persistent_Save_Manager");
-    private final static String appdataPathGameSaves = System.getenv("APPDATA") + "/Chess/save.txt";
-    private final static String appdataPathUserpreferences = System.getenv("APPDATA") + "/Chess/userPreferences.txt";
-    private final static String appdataPathUserPass = System.getenv("APPDATA") + "/Chess/Conf/C1H2E3S4S4U5S4E3R6.dat";
-    private final static String appdataFolder = System.getenv("APPDATA") + "/Chess";
-
     private final static ObjectMapper objectmapper = new ObjectMapper();
 
-    public static UserPreferences readUserprefFromSave() {
-        try {
-            BufferedReader reader = new BufferedReader(new FileReader(appdataPathUserpreferences));
-            StringBuilder sb = new StringBuilder();
-            String line = reader.readLine();
-            while (line != null) {
-                sb.append(line);
-                line = reader.readLine();
-            }
-            if (sb.toString().isEmpty()) {
-                return null;
-            } else {
-                return objectmapper.readValue(sb.toString(), UserPreferences.class);
-            }
-        } catch (FileNotFoundException e) {
-            logger.debug("No save data, creating a folder as: " + System.getenv("APPDATA") + "/Chess)");
-            File directory = new File(appdataFolder);
-            directory.mkdirs();
 
+    private final static Path gameSavesS = SavePath.getSavePath("data", "ChessGamesS.json");
+    private final static Path gameSavesU = SavePath.getSavePath("data", "ChessGamesU.json");
+    private final static Path userInfo = SavePath.getSavePath("data", "UserInfo.json");
+    private final static Path userPreferences = SavePath.getSavePath("data", "UserPref.json");
 
-        } catch (SecurityException e) {
-            logger.error("No permission to create a file! Path: " + appdataPathUserpreferences);
-        } catch (Exception e) {
-            logger.error("Error reading user pref from AppData: " + e.getMessage());
-        }
-        return null;
+    public static FileTracker<UserPreferences> userPreferenceTracker = new FileTracker<>(userPreferences.toFile(), 5, UserPreferences.getDefaultPreferences(), new TypeReference<UserPreferences>() {});
+    public static FileTracker<UserInfo> userInfoTracker = new FileTracker<>(userInfo.toFile(), 5, ChessConstants.defaultClient.info(), new TypeReference<UserInfo>() {});
+    public static FileTracker<List<GameInfo>> gameTracker = new FileTracker<>(gameSavesS.toFile(), 5, new ArrayList<>(), new TypeReference<List<GameInfo>>() {});
+    public static FileTracker<List<GameInfo>> unsavedGameTracker = new FileTracker<>(gameSavesU.toFile(), 5, new ArrayList<>(), new TypeReference<List<GameInfo>>() {});
+
+    public static List<GameInfo> getAllGames(){
+        return Stream.concat(gameTracker.getTracked().stream(), unsavedGameTracker.getTracked().stream()).collect(Collectors.toList());
     }
 
-    public static void writeUserprefToSave(UserPreferences pref) {
-        try {
-            FileWriter writer = new FileWriter(appdataPathUserpreferences, false);
-            writer.write(objectmapper.writeValueAsString(pref));
-            writer.close();
-        } catch (FileNotFoundException e) {
-            logger.debug("No save data, creating a folder as: " + System.getenv("APPDATA") + "/Chess)");
-            File directory = new File(appdataFolder);
-            directory.mkdirs();
-
-
-        } catch (SecurityException e) {
-            logger.error("No permission to create a file! Path: " + appdataPathUserpreferences);
-        } catch (Exception e) {
-            logger.error("Error writing user info to AppData: " + e.getMessage());
-        }
+    public static void removeGameFromData(UUID gameID) {
+        List<GameInfo> savedGames = gameTracker.getTracked();
+        savedGames.removeIf(game -> game.gameUUID().equals(gameID));
     }
 
-    public static FrontendClient readUserInfoFromAppData() {
-        SecretKey key = KeyManager.tryLoadSecretKey();
-        if (key == null) {
-            logger.error("Secret key could not be loaded.");
-            return null;
-        }
+    public static void updateAll(UserWGames userWGames) {
+        userPreferenceTracker.updateTracked(userWGames.user().preferences());
+        userInfoTracker.updateTracked(userWGames.user().userInfo());
+        gameTracker.updateTracked(userWGames.games());
+        unsavedGameTracker.resetTracked(); // since all coming from server, none will be unsaved
+    }
 
-        String info = readFromEncrypted(key, appdataPathUserPass);
-        if (info != null) {
+
+
+    public static class FileTracker<T> {
+        private final TypeReference<T> tTypeReference;
+        private final T defaultValue;
+        private final File file;
+
+        private T tracked;
+        private @Nullable Consumer<T> onPeriodicUpdate;
+
+        private final ReentrantReadWriteLock lock = new ReentrantReadWriteLock();
+
+        public FileTracker(File file, int updateTimeM, T defaultValue, TypeReference<T> tTypeReference) {
+            this.file = file;
+            this.defaultValue = defaultValue;
+            this.tTypeReference = tTypeReference;
+
             try {
-                return new FrontendClient(objectmapper.readValue(info, UserInfo.class));
-            } catch (Exception e) {
-                logger.error("Error mapping user info to FrontendClient object", e);
-            }
-        }
-        return null;
-    }
-
-    public static void writeUserToAppData(UserInfo info) {
-        SecretKey key = KeyManager.tryLoadSecretKey();
-        if (key == null) {
-            logger.error("Secret key could not be loaded.");
-            return;
-        }
-
-        try {
-            encryptToFile(objectmapper.writeValueAsString(info), key, appdataPathUserPass);
-        } catch (Exception e) {
-            logger.error("Error writing user info to AppData", e);
-        }
-    }
-
-    private static void encryptToFile(String data, SecretKey secretKey, String filePath) {
-        File file = new File(filePath);
-        if (!file.exists()) {
-            file.getParentFile().mkdirs();
-            logger.debug("Created new file for encryption: " + filePath);
-        }
-
-        try {
-            Cipher cipher = Cipher.getInstance("AES");
-            cipher.init(Cipher.ENCRYPT_MODE, secretKey);
-
-            try (FileOutputStream fos = new FileOutputStream(file);
-                 CipherOutputStream cos = new CipherOutputStream(fos, cipher)) {
-                cos.write(data.getBytes());
+                Files.createDirectories(file.getParentFile().toPath());
             } catch (IOException e) {
-                logger.error("IO error during file encryption at: " + filePath, e);
+                throw new RuntimeException(e);
             }
-        } catch (Exception e) {
-            logger.error("Encryption error in encryptToFile", e);
-        }
-    }
 
-    private static String readFromEncrypted(SecretKey secretKey, String filePath) {
-        File file = new File(filePath);
-        if (!file.exists()) {
-            logger.debug("No encrypted file found at: " + filePath);
+            T loaded = readValueFromSave(file);
+
+            if(loaded == null){
+                loaded = defaultValue;
+            }
+
+            this.tracked = loaded;
+
+            System.out.println("Tracked: " + loaded);
+
+            App.scheduledExecutorService.schedule(() -> {
+                forceWrite(); // Now thread-safe
+                forceUpdate(); // Safe access to tracked inside forceUpdate
+            }, updateTimeM, TimeUnit.MINUTES);
+
+            Runtime.getRuntime().addShutdownHook(new Thread(() -> {
+                forceWrite();
+                forceUpdate();
+            }));
+        }
+
+        public void setOnPeriodicUpdate(@Nullable Consumer<T> onPeriodicUpdate) {
+            this.onPeriodicUpdate = onPeriodicUpdate;
+        }
+
+        public T getTracked() {
+            lock.readLock().lock();
+            try {
+                return tracked;
+            } finally {
+                lock.readLock().unlock();
+            }
+        }
+
+        public void updateTracked(T newTracked) {
+            lock.writeLock().lock();
+            try {
+                this.tracked = newTracked;
+                System.out.println("New Tracked: " + newTracked);
+            } finally {
+                lock.writeLock().unlock();
+            }
+        }
+
+        public void resetTracked() {
+            lock.writeLock().lock();
+            try {
+                this.tracked = defaultValue;
+            } finally {
+                lock.writeLock().unlock();
+            }
+        }
+
+        public void forceWrite() {
+            lock.readLock().lock(); // Read-lock is enough if `tracked` itself is immutable or handled safely
+            try {
+                writeValueToSave(tracked, file);
+            } finally {
+                lock.readLock().unlock();
+            }
+        }
+
+        public void forceUpdate() {
+            lock.readLock().lock();
+            try {
+                if (onPeriodicUpdate != null) {
+                    onPeriodicUpdate.accept(tracked);
+                }
+            } finally {
+                lock.readLock().unlock();
+            }
+        }
+
+        private void writeValueToSave(T pass, File file) {
+            try {
+                objectmapper.writeValue(file, pass);
+            } catch (FileNotFoundException e) {
+                throw new RuntimeException(e);
+            } catch (SecurityException e) {
+                logger.error("No permission to create a file! Path: " + file);
+            } catch (Exception e) {
+                logger.error("Error writing to AppData: " + e.getMessage());
+            }
+        }
+
+        private T readValueFromSave(File file) {
+            try {
+                return objectmapper.readValue(file, tTypeReference);
+            } catch (FileNotFoundException e) {
+                logger.debug("No save data");
+            } catch (SecurityException e) {
+                logger.error("No permission to read the file! Path: " + file, e);
+            } catch (Exception e) {
+                logger.error("Error reading from AppData: ", e);
+            }
             return null;
         }
-
-        try {
-            Cipher cipher = Cipher.getInstance("AES");
-            cipher.init(Cipher.DECRYPT_MODE, secretKey);
-
-            try (FileInputStream fis = new FileInputStream(file);
-                 CipherInputStream cis = new CipherInputStream(fis, cipher)) {
-                return new String(cis.readAllBytes());
-            } catch (IOException e) {
-                logger.error("IO error during file decryption at: " + filePath, e);
-            }
-        } catch (Exception e) {
-            logger.error("Decryption error in readFromEncrypted", e);
-        }
-        return null;
     }
 
 
 
-    // game save related
-
-    public static List<ChessGame> readGamesFromAppData() {
-        List<ChessGame> games = new ArrayList<>();
-        try {
-            BufferedReader reader = new BufferedReader(new FileReader(appdataPathGameSaves));
-            String line;
-            while ((line = reader.readLine()) != null) {
-                // Check if the line contains the data to remove
-                games.add(CryptoUtils.gameFromSaveString(line));
-            }
-            reader.close();
-        } catch (FileNotFoundException e) {
-            logger.debug("No save data, creating a folder as: " + System.getenv("APPDATA") + "/Chess)");
-            File directory = new File(appdataFolder);
-            directory.mkdirs();
-
-
-        } catch (SecurityException e) {
-            logger.error("No permission to create a file! Path: " + appdataPathGameSaves);
-        } catch (Exception e) {
-            logger.error("Error reading game from AppData: " + e.getMessage());
-            e.printStackTrace();
-        }
-        return games;
-
-    }
-
-    public static void removeGameFromData(String gameHash) {
-        try {
-            // read in all games that are not the one we want to remove
-            BufferedReader reader = new BufferedReader(new FileReader(appdataPathGameSaves));
-            StringBuilder stringBuilder = new StringBuilder();
-            String line;
-            while ((line = reader.readLine()) != null) {
-                // Check if the line contains the data to remove
-                String hash = line.split(",")[0];
-                if (!hash.trim().equals(gameHash)) {
-                    // Append the line to the StringBuilder if it doesn't contain the data to remove
-                    stringBuilder.append(line).append("\n");
-                } else {
-                    logger.debug("Removing line" + line);
-
-                }
-            }
-            reader.close();
-
-            // then put all the games we want back in the file
-            BufferedWriter writer = new BufferedWriter(new FileWriter(appdataPathGameSaves, false));
-            writer.write(stringBuilder.toString());
-            writer.close();
-
-        } catch (Exception e) {
-            logger.error("Error removing game from AppData: " + e.getMessage());
-        }
-    }
-
-    public static void writeToAppData(List<ChessGame> content, boolean isAppend) {
-
-        try {
-            // Open the file in append mode
-            BufferedWriter writer = new BufferedWriter(new FileWriter(appdataPathGameSaves, isAppend));
-            for (ChessGame game : content) {
-                writer.write(CryptoUtils.chessGameToSaveString(game));
-            }
-            // Write content to the file
-
-            // Close the writer
-            writer.close();
-
-            logger.debug("Content written to the file successfully. Location: " + appdataPathGameSaves);
-        } catch (IOException e) {
-            logger.error("Error removing game from AppData: " + e.getMessage());
-        }
-    }
-
-    public static void appendGameToAppData(ChessGame game) {
-        try {
-            // Open the file in append mode
-            BufferedWriter writer = new BufferedWriter(new FileWriter(appdataPathGameSaves, true));
-            writer.write(CryptoUtils.chessGameToSaveString(game) + "\n");
-            // Write content to the file
-
-            // Close the writer
-            writer.close();
-
-            logger.debug("Content appended to the file successfully. Location: " + appdataPathGameSaves);
-        } catch (IOException e) {
-            logger.error("Error appending game to AppData: " + e.getMessage());
-        }
-    }
 }
