@@ -1,11 +1,12 @@
 package chessengine.Graphics;
 
+import chessengine.MainP.MainScreenController;
 import chessengine.Start.StartScreenController;
 
 public class DemoManager {
     private StartScreenController startController;
     private MainScreenController mainController;
-    public DemoManager(StartScreenController startController,MainScreenController mainController){
+    public DemoManager(StartScreenController startController, MainScreenController mainController){
         this.startController = startController;
         this.mainController = mainController;
 
