@@ -1,7 +1,7 @@
 package chessengine;
 
 public class Main {
-    // to trick jdk
+    // to "trick" jdk
     public static void main(String[] args) {
         App.main(args);
     }
