@@ -4,16 +4,6 @@ import chessserver.Enums.ServerResponseType;
 
 public class OutputMessage {
     ServerResponseType serverResponseType;
-
-    @Override
-    public String toString() {
-        return "OutputMessage{" +
-                "serverResponseType=" + serverResponseType +
-                ", extraInformation='" + extraInformation + '\'' +
-                ", uniqueId=" + uniqueId +
-                '}';
-    }
-
     String extraInformation;
     int uniqueId;
 
@@ -49,5 +39,14 @@ public class OutputMessage {
 
     public void setExtraInformation(String extraInformation) {
         this.extraInformation = extraInformation;
+    }
+
+    @Override
+    public String toString() {
+        return "OutputMessage{" +
+                "serverResponseType=" + serverResponseType +
+                ", extraInformation='" + extraInformation + '\'' +
+                ", uniqueId=" + uniqueId +
+                '}';
     }
 }
